# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  SignInRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Asliddin Berdiev
    url: https://github.com/asliddinberdiev
  title: EIR System API
  version: v1
paths:
  /auth/logout:
    post:
      description: Tizimdan chiqish (Sessiyani o'chirish)
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Access tokenni yangilash
      parameters:
      - description: Refresh Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
      security:
      - BearerAuth: []
      summary: Refresh Token
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Foydalanuvchi tizimga kirishi va token olishi
      parameters:
      - description: SignIn Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: SignIn
      tags:
      - auth
  /users:
    get:
      consumes:
      - application/json
      description: Fetch list of users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      security:
      - BearerAuth: []
      summary: Get users
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"